+++ ui/src/app/core/directives/longclick.directive.ts
@@ -48,8 +48,6 @@ export class LongClickDirective implements OnDestroy {
       return
     }
     this.done = false
-    event.preventDefault()
-    event.stopPropagation()
     this.downTimeout = setTimeout(() => {
       if (!this.done) {
         this.done = true