+++ CHANGELOG.md
@@ -17,6 +17,9 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 ### UI Changes
 
 - updates to the `th.json` language file (#2338) (@tomzt)
+- improvements to long-click accessory control
+  - also fixes accessory and room rearranging on the accessories page
+  - accessory and status page layout will now be locked by default on each visit
 
 ### Other Changes